Split getNumberValidator in two files each with a more specific utility
